refactor(userspace-dp): extract helpers/status, mod.rs is now a facade
Complete the #6234 decomposition: move the status-projection domain out
of server/helpers.rs into helpers/status.rs (fourth and final submodule),
leaving helpers/mod.rs as a narrow re-export facade only (module
declarations + `pub(crate) use <submodule>::*`).

status.rs owns refresh_status — the 300+-line cold-path projection onto
ProcessStatus — plus the capability/forwarding predicates it shares with
the handlers: should_run_afxdp, reconcile_status_bindings,
set_bindings_forwarding_armed, and forwarding_unsupported_error. Per the
issue's design-review amendment the first pass keeps refresh_status whole
with private zero-allocation helpers rather than exploding it into six
fragment modules; ordering and side-effect cadence are preserved exactly
(WG/GRE liveness self-heal stays gated on should_run_afxdp, the
high-cardinality neighbor-key dump stays gated on XPF_DEBUG_NEIGHBOR_KEYS,
bindings refresh before the queue/per-binding projections, worker
snapshots read before their aggregates).

Replacing the old `use super::super::*` crate-root glob with explicit
per-module imports exposed the true dependency surface the issue called
out: three crate-root imports in main.rs (afxdp::SyncedSessionEntry,
serde::Serialize, std::time::Instant) existed ONLY to feed the helpers
glob and are now dead — removed. A fourth, std::collections::BTreeMap,
is consumed only by main_tests.rs through `use super::*`, so it is
relocated to main_tests.rs as an explicit import. server/README.md is
updated to document the helpers/ submodule layout.

Pure code-motion: every moved body is byte-for-byte identical to the
pre-split source; no behavior change, so no fail-on-revert test.

Validation: the full userspace-dp bin test suite passes serially — 4121
passed, 0 failed. Targeted runs cover the moved code: server:: (89
passed, including write_state_releases_lock_before_persist #5469), the
tests:: superset (3487 passed, including the binding-plan/planning and
session-sync tests), and the three integration binaries (31 passed).
Build is green; warning count drops 176 -> 175 (dead-import cleanup, no
new warnings in helpers/). A pre-existing WG-engine concurrency test
(afxdp::wg::engine::engine_internal_tests) starves only under full
parallel test load; it passes in isolation and is untouched by this diff.

Closes #6234
