+++ userspace-dp/src/main_tests.rs
@@ -4,6 +4,7 @@
 
 use super::*;
 use crate::test_zone_ids::*;
+use std::collections::BTreeMap;
 
 fn test_zone_name_to_id() -> rustc_hash::FxHashMap<String, u16> {
     let mut m = rustc_hash::FxHashMap::default();
